Move CborDebugHelper into Internals/Cbor
Consolidates all CBOR-related code into a single folder
(Internals/Cbor) ahead of a planned CBOR/JSON debug-visibility
feature, which is being driven by community feedback that CBOR
payloads are opaque and hard to debug. CborDebugHelper previously
lived in isolation under Internals/Helpers, separate from the rest
of the CBOR converters and options in Internals/Cbor.

This is a purely mechanical move:
- No behavior change; only the namespace declaration and the
  handful of call-site `using` directives / linked-file paths in
  the Embedded.* csproj files were updated.
- CborDebugHelper is an internal type, so there is no public API
  impact.

Verified with a full solution build (0 warnings, 0 errors) and the
existing CBOR serializer test suite (97/97 passing).
